+++ lib/metro/menu.dart
@@ -17,6 +17,7 @@ All commands:
   make:stateless_widget
   make:state_managed_widget
   make:navigation_hub
+  make:journey_widget
   make:form
 
 [App Commands]