+++ lib/nylo_framework.dart
@@ -29,4 +29,4 @@ export 'package:date_field/date_field.dart';
 export 'package:dio/dio.dart';
 
 /// Nylo version
-const String nyloVersion = 'v6.7.6';
+const String nyloVersion = 'v6.8.0';